Divide Plan class
